database/config.py
import os
from pathlib import Path

from dotenv import load_dotenv
from pydantic import BaseModel
from pydantic_settings import BaseSettings

load_dotenv()

DB_NAME = os.getenv("DB_NAME")
DB_USER = os.getenv("DB_USER")
DB_PASS = os.getenv("DB_PASS")
DB_HOST = os.getenv("DB_HOST")
DB_PORT = os.getenv("DB_PORT")

DB_URL = f"postgresql+asyncpg://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"


class DBSettings(BaseModel):
    url: str = DB_URL
    echo: bool = False


class Settings(BaseSettings):
    api_v1_prefix: str = "api/v1/"
    Db_settings: DBSettings = DBSettings()


settings = Settings()
